Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 #include<math.h>
 int main(){
-    double a,b;
+    int a,b;
     scanf("%ld %ld",&a,&b);
-    if(pow(a, 2)==b&&pow(b, 2)==a){printf("Yes");}
+    if(pow(a,2)==b&&pow(b, 2)==a){printf("Yes");}
     else{printf("No");}
     return 0;
 }
